Refactor to SendBitcoin
                I've created the sendBitcoin component as a container of
                bitcoinForm interactions. Also I grouped all the Trezor
                and Ledger behavior on one container.
